# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "SuperMindsIT",
    "version": "1.0.0",
    "title": "Tapzzi API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    },
    {
      "url": "https://tapzzi-api.superminds.dev",
      "description": "Production server"
    }
    ],
  "securityDefinitions": {
      "bearerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "required": true,
        "description": "Bearer token to access these api endpoints",
        "scheme": "bearer"
      }
  }, 
  "security": [
          {
            "bearerAuth": []
          }
    ],
...
    "paths": {
      "/login": {
        "post": {
          "tags": [
            "Login"
          ],
          "summary": "Create a new token",
          "description": "This endpoint will create token",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "description": "Login Post Object",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/schemas/login"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Token Created",
              "content": {
                "application/json": {
                "schema": {
                "$ref": "#/serverResponseExample/createLoginSuccess"
                  }
                }
              }
            }, 
            "400": {
              "description": "Bad Request.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/createLoginError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        }
      },
      "/games": {
        "get": {
          "tags": [
            "Games"
          ],
          "summary": "Return Games",
          "description": "This endpoint will get all Games",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "produces": "['application/json']",
          "responses": {
            "200": {
              "description": "Games Data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/schemas/games"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Games"
          ],
          "summary": "Create a new Game",
          "description": "This endpoint will create Game",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "description": "Game Post Object",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/schemas/game"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Game Created",
              "content": {
                "application/json": {
                "schema": {
                "$ref": "#/serverResponseExample/createGameSuccess"
                  }
                }
              }
            }, 
            "400": {
              "description": "Bad Request.",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/createGameError"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        }
      },
      "/game/{id}": {
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the game that we want to match",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Games"
          ],
          "summary": "Get Game with given ID",
          "description": "This endpoint will get Game by ID.",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "Game ID",
              "schema": {
                "$ref": "#/schemas/id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Game Data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/getGameByIdSuccess"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/getGameByIdNotFoundError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "summary": "Update a Game by ID",
          "description": "This endpoint will update a game by ID",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "tags": [
            "Games"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "Update Game",
              "schema": {
                "$ref": "#/schemas/id"
              }
            }
          ],
          "requestBody": {
            "description": "Update game",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/schemas/game"
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "Game Data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/updateGameByIdSuccess"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/getGameByIdNotFoundError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete Game by ID",
          "description": "This API Endpoint will delete Game by ID",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "tags": [
            "Games"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "description": "Delete game",
              "schema": {
                "$ref": "#/schemas/id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Game Data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/deleteGameByIdSuccess"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "404": {
              "description": "Not Found Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/getGameByIdNotFoundError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        }
      },
      "/games/image": {
        "parameters": [
          {
            "name": "file",
            "in": "path",
            "required": true,
            "description": "Image that we want to upload"
          }
        ],
        "post": {
          "tags": [
            "Games"
          ],
          "summary": "Upload an Image",
          "description": "This endpoint will upload a new Image",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": { 
                  "properties": {
                    "filename": {
                      "type": "string",
                      "format": "string"
                    }
                  }
                }
              }
            }
          },
          "produces": [
            "application/json"
          ],
          "responses": {
            "201": {
              "description": "Image Game Created",
              "content": {
                "application/json": {
                "schema": {
                "$ref": "#/serverResponseExample/createGameImageSuccess"
                  }
                }
              }
            }, 
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "summary": "Delete Game Image",
          "description": "This endpoint will Delete Game Image",
          "securityDefinitions": {
            "bearerAuth": {
              "type": "apiKey",
              "in": "header",
              "name": "Authorization",
              "description": "Bearer token to access these api endpoints",
              "scheme": "bearer"
            }
          },
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "tags": [
            "Games"
          ],
          
          "responses": {
            "200": {
              "description": "Games Image Data",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/deleteGameImageByIdSuccess"
                  }
                }
              }
            },
            "401": {
              "description": "Unauthorized Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/unauthorizedError"
                  }
                }
              }
            },
            "500": {
              "description": "Server Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/serverResponseExample/serverError"
                  }
                }
              }
            }
          }
        }
      },
      "/wallpapers": {
          "get": {
            "tags": [
              "Wallpapers"
            ],
            "summary": "Returns Wallpapers",
            "description": "This endpoint will get all wallpapers",
            "security": [
              {
                "bearerAuth": []
              }
            ],
            "responses": {
              "200": {
                "description": "Wallpapers Data",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/schemas/wallpapers"
                    }
                  }
                }
              },
              "401": {
                "description": "Unauthorized Error",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/serverResponseExample/unauthorizedError"
                    }
                  }
                }
              },
              "500": {
                "description": "Server Error",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/serverResponseExample/serverError"
                    }
                  }
                }
              }
            }
          },
          "post": {
            "tags": [
              "Wallpapers"
            ],
            "summary": "Create a new Wallpaper",
            "description": "This endpoint will create new wallpaper",
            "security": [
              {
                "bearerAuth": []
              }
            ],
            "requestBody": {
              "description": "Wallpaper Request Object",
              "required": true,
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/schemas/wallpaper"
                  }
                }
              }
            },
            "produces": [
              "application/json"
            ],
            "responses": {
              "201": {
                "description": "Wallpaper Created",
                "content": {
                  "application/json": {
                  "schema": {
                  "$ref": "#/serverResponseExample/createWallpaperSuccess"
                    }
                  }
                }
              }, 
              "400": {
                "description": "Bad Request.",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/serverResponseExample/createWallpaperError"
                    }
                  }
                }
              },
              "401": {
                "description": "Unauthorized Error",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/serverResponseExample/unauthorizedError"
                    }
                  }
                }
              },
              "500": {
                "description": "Server Error",
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/serverResponseExample/serverError"
                    }
                  }
                }
              }
            }
          }
      },
...
      "/getAllData":{
        "get": {
          "tags": [
            "Default"
          ],
          "summary": "Return All Data",
          "description": "This endpoint will get all data",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "produces": "['application/json']",
          "responses": {
            "200": {
              "description": "All Data",
              "content": {
                "application/json": {
                }
              }
            }
          }
        }
      }
    },
...
  }